requirements_ingestion.txt
# Firebase → PostgreSQL Ingestion Worker
# Production dependencies for cloud deployment

# Database
psycopg2-binary>=2.9.9  # Fixed: Use compatible version for Python 3.14

# Firebase
firebase-admin==6.2.0

# Utilities (optional but recommended)
python-dotenv==1.0.0
